fix : 별점 관련 수정(23.07.21 김득렬)

+++ server/src/main/java/mainproject/cafeIn/domain/cafe/entity/Cafe.java
@@ -141,7 +141,7 @@ public class Cafe extends BaseEntity {
         }
     }
 
-    public void calculateRating(Long starRating) {
+    public void calculateRating(int starRating) {
 
         this.rating = (this.rating + starRating) / this.posts.size();
     }
